Flip dictionary in prefill implementation
Fixes a bug where prefilling the same value to more than one field
throws a duplicate key exception.

The following will now be allowed:
    { "field1": "x"
    , "field2": "x" }

The existing implementation used the resulting value as the key, and the
resulting data field as the value. This means a value can only be
assigned a single data field without triggering a "Duplicate key"
exception.

prefill.json files still use the old value/key syntax, so there are some
semantic changes. The following used to work
    { "name": "field1"
    , "email": "field1"}

And would assign (in undefined order as the dictionary is probably
unordered) first one value, and then overwrite with the other value (if
configured to overwrite values). This will now trigger a duplicate key
exception.

Closes #44