Fixing deep fs recursion issue causing IOException on non-docker linux hosts
This problem occurred because the PhysicalFileProvider with "/" as its starting point, along with 'reloadOnChange: true' will cause full filesystem recursion. On docker containers that will usually not have any adverse effects, as the user running this process has permissions to read everything. However, in a more complex environment there are files and folders in e.x. /proc where this recursion will enter, but fail with 'Invalid argument' on non-regular files belonging to other users processes.